fix issues if check missing

+++ 07_databases/mysql/functions.php
@@ -50,37 +50,43 @@
 	}
 
     function updateUser() {
-        global $connection;
+        if(isset($_POST['submit'])) {
+            global $connection;
+            $username = $_POST['username'];
+            $password = $_POST['password'];
+            $id = $_POST['id'];
 
-		$username = $_POST['username'];
-		$password = $_POST['password'];
-		$id = $_POST['id'];
+            $query = "UPDATE users SET ";
+            $query .= "username = '$username', ";
+            $query .= "password = '$password' ";
+            $query .= "WHERE id = $id ";
 
-		$query = "UPDATE users SET ";
-		$query .= "username = '$username', ";
-		$query .= "password = '$password' ";
-		$query .= "WHERE id = $id ";
+            $result = mysqli_query($connection, $query);
 
-		$result = mysqli_query($connection, $query);
-		if(!$result) {
-			die("Query FAILED ->" . mysqli_error($connection));
-		} else {
-			echo "<h5>Record UPDATED</h5>";
-		}
+            if(!$result) {
+                die("QUERY FAILED" . mysqli_error($connection));
+            } else {
+                echo "Record Updated";
+            }
+        }
     }
 
     function deleteUser() {
-        global $connection;
+	    if(isset($_POST['submit'])) {
+	        global $connection;
+	        $username = $_POST['username'];
+	        $password = $_POST['password'];
+	        $id = $_POST['id'];
 
-        $id = $_POST['id'];
+	        $query = "DELETE FROM users ";
+	        $query .= "WHERE id = $id ";
 
-        $query = "DELETE FROM users ";
-        $query .= "WHERE id = $id ";
+	        $result = mysqli_query($connection, $query);
 
-        $result = mysqli_query($connection, $query);
-        if(!$result) {
-            die("Query FAILED ->" . mysqli_error($connection));
-        } else {
-            echo "<h5>USER DELETED</h5>";
-        }
-    }
+	        if(!$result) {
+	            die("QUERY FAILED" . mysqli_error($connection));
+	        } else {
+	            echo "User Deleted";
+	        }
+	    }
+	}
